Separate source data handling from `Document`.
- Move test `util` directory to `example_output` for clarity.
